chore(useRefresh): moved folders and removed features/core to avoid future confusion
