+++ app/client/src/core/index.tsx
@@ -10,3 +10,4 @@ export * from './useIsClient';
 export * from './theme';
 export * from './useWindowDimensions';
 export * from './language';
+export * from './useRefresh';
