--- app/client/src/features/core/index.ts
@@ -1,1 +1,0 @@
-export * from './useRefresh';
